fix initdb version string parsing
The version string from 'initdb --version' produces the following on
most systems:

initdb (PostgreSQL) 14.13

However, the homebrew initdb does:

initdb (PostgreSQL) 14.15 (Homebrew)

Getting the 'last' part of the version string gives you "(Homebrew)",
not "14.15", which is what's expected in this part of the code.

Getting the third column of the version string is probably more reliable
than 'last'.